+++ backend/prisma/migrations/20250516235530_add_user_guest_relation/migration.sql
@@ -1,0 +1,14 @@
+-- CreateTable
+CREATE TABLE "Guest" (
+    "id" TEXT NOT NULL,
+    "name" TEXT NOT NULL,
+    "hostID" TEXT NOT NULL,
+    "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    "updatedAt" TIMESTAMP(3) NOT NULL,
+    "deletedAt" TIMESTAMP(3),
+
+    CONSTRAINT "Guest_pkey" PRIMARY KEY ("id")
+);
+
+-- AddForeignKey
+ALTER TABLE "Guest" ADD CONSTRAINT "Guest_hostID_fkey" FOREIGN KEY ("hostID") REFERENCES "User"("id") ON DELETE RESTRICT ON UPDATE CASCADE;