+++ backend/prisma/migrations/20250517153327_advanced_token_relations/migration.sql
@@ -1,0 +1,45 @@
+/*
+  Warnings:
+
+  - A unique constraint covering the columns `[stateId]` on the table `OAuthToken` will be added. If there are existing duplicate values, this will fail.
+  - A unique constraint covering the columns `[userId]` on the table `OAuthToken` will be added. If there are existing duplicate values, this will fail.
+  - Added the required column `stateId` to the `OAuthToken` table without a default value. This is not possible if the table is not empty.
+  - Added the required column `userId` to the `OAuthToken` table without a default value. This is not possible if the table is not empty.
+
+*/
+-- AlterTable
+ALTER TABLE "OAuthToken" ADD COLUMN     "stateId" TEXT NOT NULL,
+ADD COLUMN     "userId" TEXT NOT NULL;
+
+-- CreateTable
+CREATE TABLE "JWT" (
+    "id" TEXT NOT NULL,
+    "token" TEXT NOT NULL,
+    "validUntil" TIMESTAMP(3) NOT NULL,
+    "userId" TEXT NOT NULL,
+    "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    "updatedAt" TIMESTAMP(3) NOT NULL,
+
+    CONSTRAINT "JWT_pkey" PRIMARY KEY ("id")
+);
+
+-- CreateIndex
+CREATE UNIQUE INDEX "JWT_token_key" ON "JWT"("token");
+
+-- CreateIndex
+CREATE UNIQUE INDEX "JWT_userId_key" ON "JWT"("userId");
+
+-- CreateIndex
+CREATE UNIQUE INDEX "OAuthToken_stateId_key" ON "OAuthToken"("stateId");
+
+-- CreateIndex
+CREATE UNIQUE INDEX "OAuthToken_userId_key" ON "OAuthToken"("userId");
+
+-- AddForeignKey
+ALTER TABLE "OAuthToken" ADD CONSTRAINT "OAuthToken_stateId_fkey" FOREIGN KEY ("stateId") REFERENCES "State"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
+
+-- AddForeignKey
+ALTER TABLE "OAuthToken" ADD CONSTRAINT "OAuthToken_userId_fkey" FOREIGN KEY ("userId") REFERENCES "User"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
+
+-- AddForeignKey
+ALTER TABLE "JWT" ADD CONSTRAINT "JWT_userId_fkey" FOREIGN KEY ("userId") REFERENCES "User"("id") ON DELETE RESTRICT ON UPDATE CASCADE;