+++ backend/src/errors/authentication.ts
@@ -11,3 +11,31 @@ export class AuthenticationError extends Error {
         this.name = 'AuthenticationError';
     }
 }
+
+export class ExistingUserError extends Error {
+    constructor(message: string) {
+        super(message);
+        this.name = 'ExistingUserError';
+    }
+}
+
+export class NotVerifiedError extends Error {
+    constructor(message: string) {
+        super(message);
+        this.name = 'NotVerifiedError';
+    }
+}
+
+export class AlreadyVerifiedError extends Error {
+    constructor(message: string) {
+        super(message);
+        this.name = 'AlreadyVerifiedError';
+    }
+}
+
+export class ExpiredTokenError extends Error {
+    constructor(message: string) {
+        super(message);
+        this.name = 'ExpiredTokenError';
+    }
+}