+++ backend/src/errors/index.ts
@@ -1,1 +1,3 @@
 export * from './database.js';
+export * from './authentication.js';
+export * from './services.js';