+++ backend/src/errors/services.ts
@@ -4,3 +4,24 @@ export class AdminControlError extends Error {
         this.name = 'AdminControlError';
     }
 }
+
+export class InvalidParameterError extends Error {
+    constructor(message: string) {
+        super(message);
+        this.name = 'InvalidParameterError';
+    }
+}
+
+export class InvalidPasswordError extends Error {
+    constructor(message: string) {
+        super(message);
+        this.name = 'InvalidPasswordError';
+    }
+}
+
+export class InternalServerError extends Error {
+    constructor(message: string) {
+        super(message);
+        this.name = 'InternalServerError';
+    }
+}