+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.2-alpha.2] - 2025-05-19
+
+## Added
+
+- Own authentication and user management systems.
+
+## Changed
+
+- Adjusted the Spotify authentication, to match the new own authentication system.
+- Implemented some features in the frontend. Some not, which might break some UI features.
+
 ## [1.1.2-alpha.1] - 2025-05-17
 
 ### Changed